init: start task hooks-and-routing
