--- class-components/src/interfaces.tsx
@@ -1,7 +1,0 @@
-export interface Person {
-  name: string;
-  mass: string;
-  height: string;
-  eye_color: string;
-  gender: string;
-}
